requirements.txt
pandas>=1.5.0
psycopg2-binary>=2.9.0
SQLAlchemy>=1.4.0
openpyxl>=3.0.0
xlrd>=2.0.0
streamlit>=1.28.0
plotly>=5.15.0